feat(j-s): show files of parent case
- Moved StepFiveForm.tsx code into StepFive.tsx
